xml to json using separate document and element containers
